CHET-415: Move attachment captor to a different implementation. Remove dependency on jira-api in api
