+++ core/src/test/java/com/atlassian/migration/datacenter/core/fs/jira/listener/JiraIssueAttachmentListenerTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.atlassian.migration.datacenter.core.fs.listener;
+package com.atlassian.migration.datacenter.core.fs.jira.listener;
 
 import com.atlassian.event.api.EventPublisher;
 import com.atlassian.jira.event.issue.IssueEvent;
